chore(components): rename PreviousButton to kebab-case
It was the only PascalCase file in components/ui/.
